backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
notion-client==2.2.1
openai==1.8.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
pre-commit==3.6.0
detect-secrets==1.4.0
bandit==1.7.6
